+++ src/test/java/com/kubawach/nfs/core/model/ReceptorTestCase.java
@@ -2,7 +2,6 @@ package com.kubawach.nfs.core.model;
 
 import static org.fest.assertions.Assertions.assertThat;
 
-import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -27,7 +26,7 @@ public class ReceptorTestCase {
         // given 
         Receptor receptor = new Receptor(2L, new Threshold[]{new Threshold(null, "1", 2L)});
         State state = new State(
-                Collections.singletonMap("1", new Product(BigDecimal.valueOf(6L))), 
+                Collections.singletonMap("1", new Product(6L)), 
                 Collections.singletonMap("signal", Signal.active("1")));
         State newState = State.copy(state);
         Environment env = new Environment();
@@ -45,7 +44,7 @@ public class ReceptorTestCase {
         // given 
         Receptor receptor = new Receptor(2L, new Threshold[]{new Threshold(null, "1", 5L), new Threshold("ext-signal", "1", 7L)});
         State state = new State(
-                map(new String[] {"1", "2"}, new Product[] {new Product(BigDecimal.valueOf(8L)), new Product(BigDecimal.valueOf(0))}),
+                map(new String[] {"1", "2"}, new Product[] {new Product(8L), new Product(0)}),
                 map(new String[] {"1", "ext-signal"}, new Signal[] {Signal.charged("1", 3L), Signal.active("2")}));
         State newState = State.copy(state);
         Environment env = new Environment();
@@ -63,7 +62,7 @@ public class ReceptorTestCase {
         // given 
         Receptor receptor = new Receptor(2L, new Threshold[] {new Threshold(null, "1", 5L)});
         State state = new State(
-            Collections.singletonMap("1", new Product(BigDecimal.valueOf(6L))),
+            Collections.singletonMap("1", new Product(6L)),
             Collections.singletonMap("1", Signal.charged("1", 3L)));
         State newState = State.copy(state);
         Environment env = new Environment();
@@ -81,7 +80,7 @@ public class ReceptorTestCase {
         // given 
         Receptor receptor = new Receptor(0L, new Threshold[]{ new Threshold(null, "1", 5L)});
         State state = new State(
-            Collections.singletonMap("1", new Product(BigDecimal.valueOf(6L))),
+            Collections.singletonMap("1", new Product(6L)),
             Collections.singletonMap("1", Signal.charged("1", 0L)));
         State newState = State.copy(state);
         Environment env = new Environment();
@@ -99,7 +98,7 @@ public class ReceptorTestCase {
         // given 
         Receptor receptor = new Receptor(2L, new Threshold[]{new Threshold(null, "1", 5L), new Threshold("ext-signal", "1", 7L)});
         State state = new State(
-                map(new String[] {"1", "2"}, new Product[] {new Product(BigDecimal.valueOf(6L)), new Product(BigDecimal.valueOf(0L))}),
+                map(new String[] {"1", "2"}, new Product[] {new Product(6L), new Product(0L)}),
                 map(new String[] {"1", "ext-signal"}, new Signal[] {Signal.charged("1", 3L), Signal.active("1")}));
         State newState = State.copy(state);
         Environment env = new Environment();
